refactor: 검증 로직을 한곳에 집중
컨트롤러의 책임을 줄이기 위해 검증 로직을 한 클래스에 집중했습니다.
